Mark arguments as required
