+++ src/main/java/de/chojo/repbot/commands/channel/Channel.java
@@ -39,13 +39,13 @@ public class Channel extends SlashCommand {
                 .group(Group.of("announcement", "command.channel.announcement.description")
                         .subCommand(SubCommand.of("state", "command.channel.announcement.state.description")
                                 .handler(new State(guilds))
-                                .argument(Argument.bool("active", "command.channel.announcement.state.active.description")))
+                                .argument(Argument.bool("active", "command.channel.announcement.state.active.description").asRequired()))
                         .subCommand(SubCommand.of("where", "command.channel.announcement.where.description")
                                 .handler(new de.chojo.repbot.commands.channel.handler.announcement.Channel(guilds))
-                                .argument(Argument.channel("where", "command.channel.announcement.where.where.description")))
+                                .argument(Argument.channel("where", "command.channel.announcement.where.where.description").asRequired()))
                         .subCommand(SubCommand.of("channel", "command.channel.announcement.channel.description")
                                 .handler(new Location(guilds))
-                                .argument(Argument.channel("channel", "command.channel.announcement.channel.channel.description")))
+                                .argument(Argument.channel("channel", "command.channel.announcement.channel.channel.description").asRequired()))
                         .subCommand(SubCommand.of("info", "command.channel.announcement.info.description")
                                 .handler(new Info(guilds))))
         );
