+++ src/main/java/de/chojo/repbot/commands/channel/handler/announcement/Location.java
@@ -1,34 +1,30 @@
 package de.chojo.repbot.commands.channel.handler.announcement;
 
 import de.chojo.jdautil.interactions.slash.structure.handler.SlashHandler;
-import de.chojo.jdautil.util.Completion;
+import de.chojo.jdautil.localization.util.Replacement;
 import de.chojo.jdautil.wrapper.EventContext;
 import de.chojo.repbot.dao.provider.Guilds;
-import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
+import net.dv8tion.jda.api.entities.channel.ChannelType;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 
-public class Location implements SlashHandler {
+public class Channel implements SlashHandler {
     private final Guilds guilds;
 
-    public Location(Guilds guilds) {
+    public Channel(Guilds guilds) {
         this.guilds = guilds;
     }
 
     @Override
     public void onSlashCommand(SlashCommandInteractionEvent event, EventContext context) {
         var announcements = guilds.guild(event.getGuild()).settings().announcements();
-        if (announcements.sameChannel("same channel".equalsIgnoreCase(event.getOption("where", OptionMapping::getAsString)))) {
-            event.reply(context.localize("command.channel.announcement.location.message.samechannel")).queue();
-        } else {
-            event.reply(context.localize("command.channel.announcement.location.message.otherchannel")).queue();
+        var channel = event.getOption("channel");
+        if (channel.getChannelType() != ChannelType.TEXT) {
+            event.reply(context.localize("error.onlyTextChannel")).setEphemeral(true).queue();
+            return;
         }
-    }
 
-    @Override
-    public void onAutoComplete(CommandAutoCompleteInteractionEvent event, EventContext context) {
-        if ("where".equals(event.getFocusedOption().getName())) {
-            event.replyChoices(Completion.complete("", "same channel", "custom channel")).queue();
-        }
+        announcements.channel(channel.getAsChannel().asTextChannel());
+        event.reply(context.localize("command.channel.announcement.channel.message.set",
+                Replacement.createMention(channel.getAsChannel().asTextChannel()))).queue();
     }
 }
